[26/02/26] Add & Fix
Add
크리티컬 확률에 따른 데미지 계산 추가
- 발사할 때 FMath::RandRange로 0~1 사이의 값을 IsCritValue로 정함.
- IsCritValue가 현재 크리티컬 확률 이하인 경우 크리티컬 데미지가 발생하도록 하는 방식.

Fix
크리티컬 확률, 크리티컬 데미지가 int32로 되어있던 부분들을 float으로 수정.
- 크리티컬 확률 : 0 ~ 1 사이의 값이 되어야하기 때문
- 크리티컬 데미지 : Multiplier이기 때문에 계수처럼 현재 데미지에 1.5배, 1.7배를 곱해줘야하는 부분이라서

+++ Source/TheSeventhbullet/Data/StatusTypes.h
@@ -92,10 +92,10 @@ struct FCharacterStat
 	int32 Defence;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Stat")
-	int32 CriticalChance;//크리 확률
+	float CriticalChance;//크리 확률
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Stat")
-	int32 CriticalDamage;//크피 확률
+	float CriticalDamage;//크피 확률
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Stat")
 	int32 Speed;
@@ -107,8 +107,8 @@ struct FCharacterStat
 		Stamina = 0;
 		Attack = 0;
 		Defence = 0;
-		CriticalChance = 0;
-		CriticalDamage = 0;
+		CriticalChance = 0.f;
+		CriticalDamage = 0.f;
 		Speed = 0.0f;
 	}
